Assorted tweaks and cleanups

+++ network_simulation/experiment.py
@@ -18,22 +18,26 @@ class Experiment:
             user_input = input("Enter 'quit' or 'exit' to stop the experiment: ").strip().lower()
             if user_input in {"quit", "exit"}:
                 self.termination_flag = True
-                print("Termination signal received. Stopping experiment...")
+                print("Termination signal received. Slow-stopping experiment...")
                 break
 
     def is_scenario_completed(self, scenario_dir, expected_num_rows):
         """Check if the scenario is completed based on existing metrics."""
-        metrics_path = os.path.join(scenario_dir, "summary_metrics.csv")
-        if os.path.exists(metrics_path):
-            try:
-                metrics_df = pd.read_csv(metrics_path)
-                if len(metrics_df) >= expected_num_rows:    # if the metrics file is long enough
-                    print(f"Skipping completed scenario: {scenario_dir}")
-                    return True
-                else:
-                    shutil.rmtree(scenario_dir)
-            except Exception as e:
-                print(f"Error reading metrics file for {scenario_dir}: {e}")
+        if not os.path.exists(scenario_dir):
+            return False
+        for file_name in os.listdir(scenario_dir):
+            if file_name.startswith("summary_metrics") and file_name.endswith(".csv"):
+                metrics_path = os.path.join(scenario_dir, file_name)
+                if os.path.exists(metrics_path):
+                    try:
+                        metrics_df = pd.read_csv(metrics_path)
+                        if len(metrics_df) >= expected_num_rows:    # if the metrics file is long enough
+                            print(f"Skipping completed scenario: {scenario_dir}")
+                            return True
+                        else:
+                            shutil.rmtree(scenario_dir)
+                    except Exception as e:
+                        print(f"Error reading metrics file for {scenario_dir}: {e}")
         return False
 
     def run_simulation(self, num_nodes, num_connections, output_dir, num_steps, display_interval, metrics_interval, random_seed, color_by=ColorBy.ACTIVITY):
@@ -41,11 +45,11 @@ class Experiment:
         sim = Simulation(num_nodes=num_nodes, num_connections=num_connections, output_dir=output_dir, random_seed=random_seed)
         if self.termination_flag:
             return f"Simulation {random_seed, num_nodes, num_connections} terminated by user."
-        print(f"Simulation starting for {num_connections} connections with seed {random_seed}")
+        print(f"Simulation starting for {num_nodes} nodes, {num_connections} connections with seed {random_seed}")
         sim.run(num_steps=num_steps, display_interval=display_interval, metrics_interval=metrics_interval, show=False, color_by=color_by)
-        return f"Simulation completed for {num_connections} connections with seed {random_seed}"
+        return f"Simulation completed for {num_nodes} nodes, {num_connections} connections with seed {random_seed}"
 
-    def run_experiment(self, seed_range, nodes_range, connections_range, num_steps, display_interval, metrics_interval):
+    def run_experiment(self, seed_range, nodes_range, connections_range, num_steps, display_interval, metrics_interval, connections_as_density=False):
         # Start thread to check for early termination
         threading.Thread(target=self.monitor_input_early_termination, daemon=True).start()
 
@@ -54,10 +58,13 @@ class Experiment:
         with ProcessPoolExecutor(max_workers=os.cpu_count()) as executor:
             futures = []
             for num_nodes, num_connections, seed in itertools.product(nodes_range, connections_range, seed_range):
+                if connections_as_density:   # decimal values for num_connections represent network density, and are converted
+                    num_connections = int(num_connections * (num_nodes * (num_nodes - 1) / 2))
                 scenario_dir = os.path.join(self.experiment_folder, f"seed_{seed}", f"nodes_{num_nodes}", f"edges_{num_connections}")
                 if self.is_scenario_completed(scenario_dir, num_steps / metrics_interval):
                     continue
 
+                # Add simulation to queue
                 futures.append(
                     executor.submit(
                         self.run_simulation,
